archive/n8n_betflow/betflow-pro/requirements.txt
notion-client==2.0.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
APScheduler==3.10.1
requests-cache==1.1.0
pytz==2023.3

